More application flow added
More progress made in the overall program. Mostly all that it still lacks is the proper verification of text fields that were used and ensuring that the user password and username matches where relevant. Also might be a good idea to Block non-ideal text elements. Also might be a good idea to do a password hash to the db. This will help use ensure passwords are safe - in a mundane sense.